feat: plugins working (again!)
