+++ src/modules/plugins/homepage-switcher.ts
@@ -1,16 +1,13 @@
 export default async function defineWxtPlugin() {
-  defineStPlugin(
-    "homepage-switcher",
-    () => {
-      let logos = Array.from(document.getElementsByClassName("logo")) as HTMLAnchorElement[];
-      logos.forEach((logo) => {
-        logo.addEventListener("click", async function (e) {
-          if (window.location.pathname === "/") return;
-          e.preventDefault();
-          let tab = logos[0].href;
-          browser.runtime.sendMessage({ toTab: tab });
-        });
+  defineStPlugin("homepage-switcher", () => {
+    let logos = Array.from(document.getElementsByClassName("logo")) as HTMLAnchorElement[];
+    logos.forEach((logo) => {
+      logo.addEventListener("click", async function (e) {
+        if (window.location.pathname === "/") return;
+        e.preventDefault();
+        let tab = logos[0].href;
+        browser.runtime.sendMessage({ toTab: tab });
       });
-    }
-  );
+    });
+  });
 }
