+++ src/modules/plugins/legacy-timetable.ts
@@ -1,19 +1,16 @@
 export default async function defineWxtPlugin() {
-  defineStPlugin(
-    "legacy-timetable",
-    () => {
-      if (window.location.pathname === "/" && document.querySelector(".timetable")) {
-        // get the timetable container and add the 'columns' class
-        let timetableContainer = document.querySelector("[data-timetable-container]");
-        if (timetableContainer) {
-          timetableContainer.classList.add("columns");
-          // get the row element and move the timetable container to the beginning
-          let rowElement = document.querySelector(".Component_Dashboard_TimetableController")?.parentNode?.parentNode;
-          if (rowElement) {
-            rowElement.insertBefore(timetableContainer, rowElement.firstChild);
-          }
+  defineStPlugin("legacy-timetable", () => {
+    if (window.location.pathname === "/" && document.querySelector(".timetable")) {
+      // get the timetable container and add the 'columns' class
+      let timetableContainer = document.querySelector("[data-timetable-container]");
+      if (timetableContainer) {
+        timetableContainer.classList.add("columns");
+        // get the row element and move the timetable container to the beginning
+        let rowElement = document.querySelector(".Component_Dashboard_TimetableController")?.parentNode?.parentNode;
+        if (rowElement) {
+          rowElement.insertBefore(timetableContainer, rowElement.firstChild);
         }
       }
     }
-  );
+  });
 }
