+++ src/modules/plugins/scroll-period.js
@@ -1,24 +1,26 @@
 export default defineWxtPlugin(() => {
-  const timetable = document.querySelector("[data-timetable-container] div.scrollable");
+  defineStPlugin("scroll-period", () => {
+    const timetable = document.querySelector("[data-timetable-container] div.scrollable");
 
-  if (window.location.pathname === "/" && document.getElementsByClassName("timetable")[0]) {
-    setTimeout(updateScrollbar, 500);
-    setInterval(updateScrollbar, 10000);
-  }
+    if (window.location.pathname === "/" && document.getElementsByClassName("timetable")[0]) {
+      setTimeout(updateScrollbar, 500);
+      setInterval(updateScrollbar, 10000);
+    }
 
-  function updateScrollbar() {
-    const currentPeriod = getCurrentPeriod();
-    if (currentPeriod && currentPeriod.index) {
-      // console.log("scrolling to period", periodIndex);
-      try {
-        const period = document.querySelector(`.timetable thead tr th:nth-child(${periodIndex})`);
-        timetable.scroll({
-          left: period.offsetLeft - 55, // 16 for perfect alignment
-          behavior: "smooth", // or 'auto' to scroll instantly
-        });
-      } catch (error) {
-        console.error("Error updating scrollbar:", error);
+    function updateScrollbar() {
+      const currentPeriod = getCurrentPeriod();
+      if (currentPeriod && currentPeriod.index) {
+        // console.log("scrolling to period", periodIndex);
+        try {
+          const period = document.querySelector(`.timetable thead tr th:nth-child(${periodIndex})`);
+          timetable.scroll({
+            left: period.offsetLeft - 55, // 16 for perfect alignment
+            behavior: "smooth", // or 'auto' to scroll instantly
+          });
+        } catch (error) {
+          console.error("Error updating scrollbar:", error);
+        }
       }
     }
-  }
+  });
 });
