+++ src/modules/plugins/scroll-segments/index.ts
@@ -1,0 +1,15 @@
+import styleText from "./scroll-segments.css?inline";
+
+export default defineWxtPlugin(() => {
+  defineStPlugin(
+    "scroll-segments",
+    () => {
+      const content = document.getElementById("content");
+      const footer = document.getElementById("footer");
+      if (content && footer) {
+        content.appendChild(footer);
+      }
+      injectStyles(styleText);
+    }
+  );
+});
