+++ src/modules/plugins/scroll-segments/scroll-segments.css
@@ -10,7 +10,7 @@
 }
 
 body {
-  overflow-y: hidden;
+  overflow-y: hidden !important;
 }
 
 #footer {
